pyproject.toml
[tool.poetry]
name = "youtrack-worktime-reporter"
version = "0.1.0"
description = ""
authors = ["Vladislav Bortnikov <bortnikov.vladislav@e-sakha.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
youtrack = "^0.1.69"
python-dotenv = "^0.14.0"
influxdb = "^5.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
